refactor(components): consolidate three error boundaries into one
Merge ScreenErrorBoundary and RetryErrorBoundary into a single prop-driven
ErrorBoundary at src/components/common. The merged component keeps the
original reporting behaviour (Sentry with an errorBoundary boundary tag plus
crashReportingService and the logger) and adds optional auto-retry with
exponential backoff (autoRetry, maxRetries, isTransient, onRetrySuccess,
onMaxRetriesReached) and route-key auto-reset (resetKeys), both opt-in so the
existing ~20 boundary consumers are unchanged. app/_layout.tsx now uses one
boundary instead of the double ErrorBoundary + RetryErrorBoundary nesting.
Delete the superseded ScreenErrorBoundary, RetryErrorBoundary and
DefaultErrorFallback modules and their duplicated tests, consolidate the
coverage into one ErrorBoundary test, and update the strategy doc.
