+++ src/components/index.ts
@@ -4,7 +4,7 @@ export * from './common/AppText';
 export * from './common/AccessibleModal';
 export * from './common/ModalPortal';
 export { ErrorBoundary } from './common/ErrorBoundary';
-export type { ErrorBoundaryFallbackProps } from './common/ErrorBoundary';
+export type { ErrorBoundaryFallbackProps, ErrorBoundaryProps } from './common/ErrorBoundary';
 export { KeyboardDelegateProvider, useKeyboardState } from './common/KeyboardDelegateProvider';
 export { default as PrimaryButton } from './common/PrimaryButton';
